removes parse logic from deserialize/mod
